# openapi.yml
openapi: 3.0.3
info:
  title: Family Recipes
  description: |-
    The purpose of Family Recipes is to make sharing recipes easier. Instead of asking for Grandma’s old chocolate chip cookie recipe that has been kept on the same piece of paper for decades, users can save and share recipes with their families and other groups from Family Recipes. Users are also able to view recipes shared with them from various groups in a format similar to social media. Family Recipes allows users to easily see recipes recommended by people they know and share their favorite recipes with others.
  version: 1.0.0
servers:
  - url: TBD
...
paths:
...
  /post:
    post:
      tags:
        - post
      summary: Create post
      description: Create a post
      operationId: createPost
      requestBody:
        $ref: '#/components/requestBodies/createPost'
      responses:
        '201':
          description: Successfully created post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/createRes'
        '400':
          description: Bad request
  /post/{postId}:
    parameters:
      - name: postId
        in: path
        description: ID of post to get
        required: true
        schema:
          type: integer
          format: int64
      - name: embedRecipe
        in: query
        description: Whether or not to include recipe in response, default true
        required: false
        schema:
            type: boolean
    get:
      tags:
        - post
      summary: Get post
      description: Get single post
      operationId: getSinglePost
      responses:
        '200':
          description: Successfully got post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/singlePostRes'
        '404':
          description: Not found
    delete:
      tags:
        - post
      summary: Delete post
      description: Delete a post
      operationId: deleteSinglePost
      responses:
        '200':
          description: Successfully deleted post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/deleteRes'
        '404':
          description: Not found
  /post/user/{username}:
    parameters:
      - name: username
        in: path
        description: Username of user to get posts from
        required: true
        schema:
          type: string
      - name: embedRecipe
        in: query
        description: Whether or not to include recipes in response, default true
        required: false
        schema:
            type: boolean
    get:
      tags:
        - post
      summary: Get posts by user
      description: Get posts by user
      operationId: getUserPosts
      responses:
        '200':
          description: Successfully got posts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/multiPostRes'
        '404':
          description: Not found
  /post/group/{groupId}:
    parameters:
      - name: groupId
        in: path
        description: ID of group to get posts from
        required: true
        schema:
          type: integer
          format: int64
      - name: embedRecipe
        in: query
        description: Whether or not to include recipes in response, default true
        required: false
        schema:
            type: boolean
    get:
      tags:
        - post
      summary: Get posts by group
      description: Get posts by group
      operationId: getGroupPosts
      responses:
        '200':
          description: Successfully got posts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/multiPostRes'
        '404':
          description: Not found
...
components:
  schemas:
...
    recipe:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1234
        title:
          type: string
          example: Mac and Cheese
        description:
          type: string
          example: Delicious Mac and Cheese
        ingredients:
          type: array
          items:
            type: string
            description: Ingredients needed for recipe
        instructions:
          type: string
          example: Make the mac and cheese
        difficultyLevel:
          type: string
          description: Level of difficulty
          example: Easy
          enum:
            - Easy
            - Medium
            - Hard
            - Expert
        image:
          type: string
          example: Base64EncodedString
        foodType:
          type: string
          example: Pasta
        recipeSource:
          type: string
...
    createRes:
      type: object
      properties:
        status:
          type: integer
          example: 200
        id:
          type: integer
          example: 1234
    deleteRes:
      type: object
      properties:
        status:
          type: integer
          example: 200
        rowsAffected:
          type: integer
          example: 1
    singlePostRes:
      type: object
      properties:
        status:
          type: integer
          example: 200
        data:
          type: object
          properties:
            id:
              type: integer
              format: int64
              example: 1234
            author:
              type: string
              example: user1
            dateCreated:
              type: string
              format: date-time
            likes:
              type: integer
              format: int64
              description: Number of users who have liked the post
            recipeId:
              type: integer
              format: int64
              description: Id of recipe featured in post
              example: 1234
            tags:
              type: array
              items:
                type: string
                description: Tags to associate with the post
            recipe:
              $ref: "#/components/schemas/recipe"
    multiPostRes:
      type: object
      properties:
        status:
          type: integer
          example: 200
        data:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
                format: int64
                example: 1234
              author:
                type: string
                example: user1
              dateCreated:
                type: string
                format: date-time
              likes:
                type: integer
                format: int64
                description: Number of users who have liked the post
              recipeId:
                type: integer
                format: int64
                description: Id of recipe featured in post
                example: 1234
              tags:
                type: array
                items:
                  type: string
                  description: Tags to associate with the post
              recipe:
                $ref: "#/components/schemas/recipe"
...
  requestBodies:
...
    createPost:
      content:
        application/json:
          schema:
            type: object
            properties:
              authorId:
                type: integer
                format: int64
                example: 1234
              recipeId:
                type: integer
                format: int64
                example: 1234
              tags:
                type: array
                items:
                  type: string
                  description: Tags to associate with the post
...